Time: 0 ms (100.00%) | Memory: 8.1 MB (7.20%) - LeetSync

+++ 171-excel-sheet-column-number/excel-sheet-column-number.cpp
@@ -6,7 +6,7 @@ public:
 
         for(int i : s){
             if(size == -1) break;
-            long c = i - 'A' +1;
+            int c = i - 'A' +1;
             long power = pow(26,size);
             result += (c*power);
             size--;
